+++ R/tm_g_heat_bygrade.R
@@ -214,6 +214,14 @@ ui_g_heatmap_bygrade <- function(id, ...) {
       plot_decorate_output(id = ns(NULL))
     ),
     encoding = div(
+      ### Reporter
+      shiny::tags$div(
+        teal.reporter::add_card_button_ui(ns("addReportCard")),
+        teal.reporter::download_report_button_ui(ns("downloadButton")),
+        teal.reporter::reset_report_button_ui(ns("resetButton"))
+      ),
+      shiny::tags$br(),
+      ###
       teal.widgets::optionalSelectInput(
         ns("id_var"),
         "ID Variable",
@@ -288,6 +296,7 @@ ui_g_heatmap_bygrade <- function(id, ...) {
 
 srv_g_heatmap_bygrade <- function(id,
                                   datasets,
+                                  reporter,
                                   sl_dataname,
                                   ex_dataname,
                                   ae_dataname,
@@ -295,9 +304,13 @@ srv_g_heatmap_bygrade <- function(id,
                                   label,
                                   plot_height,
                                   plot_width) {
+  with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
+
   moduleServer(id, function(input, output, session) {
     teal.code::init_chunks()
-    font_size <- srv_g_decorate(id = NULL, plt = plt, plot_height = plot_height, plot_width = plot_width) # nolint
+    decorate_output <- srv_g_decorate(id = NULL, plt = plt, plot_height = plot_height, plot_width = plot_width) # nolint
+    font_size <- decorate_output$font_size
+    pws <- decorate_output$pws
 
     observeEvent(cm_dataname, {
       if (!is.na(cm_dataname)) {
@@ -443,5 +456,27 @@ srv_g_heatmap_bygrade <- function(id,
       modal_title = paste("R code for", label),
       datanames = datasets$datanames()
     )
+
+    ### REPORTER
+    if (with_reporter) {
+      card_fun <- function(comment) {
+        card <- teal.reporter::TealReportCard$new()
+        card$set_name("Heatmap by Grade")
+        card$append_text("Heatmap by Grade", "header2")
+        card$append_text("Filter State", "header3")
+        card$append_fs(datasets$get_filter_state())
+        card$append_text("Plot", "header3")
+        card$append_plot(plt(), dim = pws$dim())
+        if (!comment == "") {
+          card$append_text("Comment", "header3")
+          card$append_text(comment)
+        }
+        card
+      }
+
+      teal.reporter::add_card_button_srv("addReportCard", reporter = reporter, card_fun = card_fun)
+      teal.reporter::download_report_button_srv("downloadButton", reporter = reporter)
+      teal.reporter::reset_report_button_srv("resetButton", reporter)
+    }
   })
 }